requirements.txt
torch
torchvision
torchaudio

tqdm>=4.64
scikit-learn>=1.3
pyyaml>=6.0
matplotlib>=3.7
seaborn>=0.12
